+++ CHANGELOG.md
@@ -1,12 +1,18 @@
 # HEAD
 
-# May 21, 2025
+## 2025-05-21
+-   Implement demo command [PR#8](https://github.com/matteoilardi/Raytracer/pull/8)
+-   Create shapes.hpp library [PR#8](https://github.com/matteoilardi/Raytracer/pull/8)
+-   Import external library CLI11 [PR#8](https://github.com/matteoilardi/Raytracer/pull/8)
 -   Implement workflow for continuous integration [PR#6](https://github.com/matteoilardi/Raytracer/pull/6)
 -   Implement Google Test unit testing framework [PR#6](https://github.com/matteoilardi/Raytracer/pull/6)
 
-# May 2, 2025
+## 2025-05-02
 -   Fix an issue with the vertical order of the images [#5](https://github.com/matteoilardi/Raytracer/pull/5)
 
-# Version 0.1.0
 
+#   [0.1.0] - 2025-04-13
 -   First release of the code
+-   Create main executable implementing PFM to PNG image conversion
+-   Create colors.hpp, geometry.hpp and cameras.hpp libraries
+-   Import external library stb_image_write.h